+++ src/features/rename.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode'
 import { symbolTables } from './syntax_errors';
-import { DeclarationType, TamarinSymbol, TamarinSymbolTable } from '../symbol_table/create_symbol_table';
+import { CreateSymbolTableResult, DeclarationType, TamarinSymbol, TamarinSymbolTable } from '../symbol_table/create_symbol_table';
 
 export function get_symbol(range : vscode.Range, symbol_table : TamarinSymbolTable): TamarinSymbol|null{
     for (let k = 0 ; k < symbol_table.getSymbols().length; k++){
@@ -11,7 +11,8 @@ export function get_symbol(range : vscode.Range, symbol_table : TamarinSymbolTab
     return null;
 }
 
-async function replace_symbol(symbol : TamarinSymbol, editor : vscode.TextEditor, newName : string ){
+/* Function used to replace a symbol by the user choice */ 
+ function replace_symbol(symbol : TamarinSymbol, editor : vscode.TextEditor, newName : string, edit : vscode.WorkspaceEdit): void{
   if (symbol.name === undefined) {
     console.error("Cannot replace symbol with undefined name");
     return;
@@ -23,25 +24,17 @@ async function replace_symbol(symbol : TamarinSymbol, editor : vscode.TextEditor
   const newNameLength = newName.length;
   const newRange = new vscode.Range(
     range.start,
-    editor.document.positionAt(editor.document.offsetAt(range.start) + newNameLength)
+    editor.document.positionAt(editor.document.offsetAt(range.start) + newNameLength - oldNameLength)
   );
-  const edit = new vscode.WorkspaceEdit();
-  edit.replace(editor.document.uri, range, newText);
-  await vscode.workspace.applyEdit(edit);
   symbol.name_range = newRange;
   symbol.name = newName;
-  // Mettre à jour la position de fin du symbole en utilisant la longueur du nouveau nom et la longueur de l'ancien nom
-  const delta = newNameLength - oldNameLength;
-  symbol.name_range = new vscode.Range(
-    symbol.name_range.start,
-    symbol.name_range.end.translate(0, delta)
-  );
+  edit.replace(editor.document.uri, range, newText);
 }
 
 
 export function RenameCommand(context : vscode.ExtensionContext){
 
-    const  provideRenameCodeAction = vscode.commands.registerCommand("Rename", async () => {
+    const provideRenameCodeAction = vscode.commands.registerCommand("Rename", async () => {
         const activeEditor = vscode.window.activeTextEditor;
         if (!activeEditor) {
             vscode.window.showErrorMessage("No active text editor");
@@ -50,7 +43,7 @@ export function RenameCommand(context : vscode.ExtensionContext){
 
         const activeFile = activeEditor.document.uri.path.split('/').pop(); 
         if(activeFile){
-            const symbolTable = symbolTables.get(activeFile);
+            const symbolTable : CreateSymbolTableResult|undefined = symbolTables.get(activeFile);
             if (!symbolTable) {
                 vscode.window.showErrorMessage("No symbol table for this file");
                 return;
@@ -77,8 +70,6 @@ export function RenameCommand(context : vscode.ExtensionContext){
         const range = new vscode.Range(startPos, endPos);
         const corresponding_symbol = get_symbol(range, symbolTable.symbolTable)
 
-        
-
         const newName = await vscode.window.showInputBox({
             placeHolder: "Enter the new name"
         });
@@ -86,34 +77,37 @@ export function RenameCommand(context : vscode.ExtensionContext){
             vscode.window.showErrorMessage("No new name provided");
             return;
         }
-
-
         let replace_count: number = 0;
+        const edit = new vscode.WorkspaceEdit();
+        const chosenSymbolName = corresponding_symbol?.name;
+        
         for(let k = 0 ; k < symbolTable.symbolTable.getSymbols().length ; k++){
             let current_symbol = symbolTable.symbolTable.getSymbol(k);
             if(current_symbol.declaration === DeclarationType.PRVariable || current_symbol.declaration === DeclarationType.ActionFVariable || current_symbol.declaration === DeclarationType.CCLVariable){
-                if(corresponding_symbol?.context === current_symbol.context && corresponding_symbol?.name === current_symbol.name){
+                if(corresponding_symbol?.context === current_symbol.context && chosenSymbolName === current_symbol.name){
                     replace_count ++
-                    replace_symbol(current_symbol, activeEditor, newName);
-                } 
-            }
-            else if (current_symbol.declaration === DeclarationType.LemmaVariable){
-                if(current_symbol.associated_qf?.id === corresponding_symbol?.associated_qf?.id && current_symbol.name === corresponding_symbol?.name){
-                    replace_count ++
-                    replace_symbol(current_symbol, activeEditor, newName);
+                    replace_symbol(current_symbol, activeEditor, newName, edit);
                 }
             }
-            else if( current_symbol.name === corresponding_symbol?.name && current_symbol.declaration === corresponding_symbol?.declaration){
+            else if (current_symbol.declaration === DeclarationType.LemmaVariable){
+                if(current_symbol.associated_qf?.id === corresponding_symbol?.associated_qf?.id && current_symbol.name === chosenSymbolName){
+                    replace_count ++
+                    replace_symbol(current_symbol, activeEditor, newName, edit);
+                }
+            }
+            else if( current_symbol.name === chosenSymbolName && current_symbol.declaration === corresponding_symbol?.declaration){
                 replace_count ++
-                replace_symbol(current_symbol, activeEditor, newName);
+                replace_symbol(current_symbol, activeEditor, newName, edit);
             }
         }
+
+        vscode.workspace.applyEdit(edit);
 
         if(replace_count === 0){
             vscode.window.showErrorMessage("No such token");
         }
         
-        }
+    }
     });
 
     context.subscriptions.push(provideRenameCodeAction);