+++ src/features/wellformedness_checks.ts
@@ -6,13 +6,14 @@ import { getName } from './syntax_errors';
 
 
 
-function build_error_display(node : Parser.SyntaxNode, editeur: vscode.TextEditor, diags : vscode.Diagnostic[], message : string){
+function build_error_display(node : Parser.SyntaxNode, editeur: vscode.TextEditor, diags : vscode.Diagnostic[], message : string): vscode.Diagnostic{
     let start = node.startIndex;
     let end  = node.endIndex;
     let positionStart = editeur.document.positionAt(start);
     let positionEnd = editeur.document.positionAt(end);
     let diag = new vscode.Diagnostic(new vscode.Range(positionStart, positionEnd ), message, vscode.DiagnosticSeverity.Error);
-    diags.push(diag)
+    diags.push(diag);
+    return diag;
 }
 
 function build_warning_display(node : Parser.SyntaxNode, editeur: vscode.TextEditor, diags : vscode.Diagnostic[], message : string): vscode.Diagnostic{
@@ -21,13 +22,14 @@ function build_warning_display(node : Parser.SyntaxNode, editeur: vscode.TextEdi
     let positionStart = editeur.document.positionAt(start);
     let positionEnd = editeur.document.positionAt(end);
     let diag = new vscode.Diagnostic(new vscode.Range(positionStart, positionEnd ), message, vscode.DiagnosticSeverity.Warning);
-    diags.push(diag)
-    return diag
+    diags.push(diag);
+    return diag;
 }
 
 
-
-function levenshteinDistance(s1: string , s2 : string) {
+/* Function used to compare the distance between two strings,
+ returns the minimum operations required to convert the first string into the second one */
+function levenshteinDistance(s1: string , s2 : string): number {
     const matrix = [];
 
     // increment along the first column of each row
@@ -60,6 +62,7 @@ function levenshteinDistance(s1: string , s2 : string) {
     return matrix[s2.length][s1.length];
 }
 
+// Function to get the grammar type of a node 
 function get_child_grammar_type(node :Parser.SyntaxNode): string[]{
     let results : string[] = []
     for(let child of node.children) {
@@ -68,7 +71,9 @@ function get_child_grammar_type(node :Parser.SyntaxNode): string[]{
     return results;
 }
 
-export function check_reserved_facts(node : Parser.SyntaxNode, editor : vscode.TextEditor, diags : vscode.Diagnostic[]){
+/* Function used to perform checks on all reserved facts,
+ checks if they are in the wright place and used with the correct arity */
+export function check_reserved_facts(node : Parser.SyntaxNode, editor : vscode.TextEditor, diags : vscode.Diagnostic[]): void{
     for(let child of node.children){
         if(child.grammarType === DeclarationType.LinearF ||child.grammarType === DeclarationType.PersistentF){
             const fact_name = getName(child.child(0), editor);
@@ -99,6 +104,22 @@ export function check_reserved_facts(node : Parser.SyntaxNode, editor : vscode.T
             else if((fact_name === ReservedFacts[3] || fact_name === ReservedFacts[4] || fact_name === ReservedFacts[5]) && node.parent?.grammarType === 'simple_rule' ){
                 build_warning_display(child, editor, diags,  "You are not supposed to use KD KU or action K in a rule ");
             }
+            else if( fact_name === ReservedFacts[6]){
+                if(get_arity(child.child(2)?.children) != 2){
+                    build_error_display(child, editor, diags, "Error : incorrect arity for diff fact, 2 arguments expected")
+                }
+            }
+        }
+        else if (child.grammarType === 'nary_app'){
+            const fact_name_node = child.child(0);
+            if (fact_name_node !== null && fact_name_node !== undefined) {
+                const fact_name = getName(fact_name_node, editor);
+                if(fact_name === ReservedFacts[6]){
+                    if(node.grammarType === DeclarationType.Equation || node.grammarType === 'mset_term'){
+                        build_warning_display(child, editor, diags , "Warning  :  diff fact cannot be used in an equation")
+                    }
+                }
+            }
         }
         else {
             check_reserved_facts(child, editor, diags)
@@ -107,18 +128,21 @@ export function check_reserved_facts(node : Parser.SyntaxNode, editor : vscode.T
     
 };
 
+/* Function used to check if all variables with the same name  in a rule have the same type,
+Also checks that if a variables is in the right side of a macro or equation it is also present in the left side  */
 //A optimiser peut être 
-function check_variables_type_is_consistent_inside_a_rule(symbol_table : TamarinSymbolTable, editor: vscode.TextEditor, diags:vscode.Diagnostic[]){
+function check_variables_type_is_consistent_inside_a_rule(symbol_table : TamarinSymbolTable, editor: vscode.TextEditor, diags:vscode.Diagnostic[]) : void{
     for (let i = 0 ; i < symbol_table.getSymbols().length; i++){
-        if(symbol_table.getSymbol(i).declaration === DeclarationType.PRVariable || symbol_table.getSymbol(i).declaration === DeclarationType.CCLVariable || symbol_table.getSymbol(i).declaration === DeclarationType.ActionFVariable ){
+        let current_symbol = symbol_table.getSymbol(i);
+        if(current_symbol.declaration === DeclarationType.PRVariable || current_symbol.declaration === DeclarationType.CCLVariable || current_symbol.declaration === DeclarationType.ActionFVariable ){
             for(let j = 0; j < symbol_table.getSymbols().length; j++){
-                if((symbol_table.getSymbol(i).declaration === DeclarationType.PRVariable || symbol_table.getSymbol(i).declaration === DeclarationType.CCLVariable || symbol_table.getSymbol(i).declaration === DeclarationType.ActionFVariable) && i !== j){
-                    if(symbol_table.getSymbol(i).context === symbol_table.getSymbol(j).context && symbol_table.getSymbol(i).name === symbol_table.getSymbol(j).name){
-                        if(symbol_table.getSymbol(i).type === symbol_table.getSymbol(j).type){
+                if((current_symbol.declaration === DeclarationType.PRVariable || current_symbol.declaration === DeclarationType.CCLVariable || current_symbol.declaration === DeclarationType.ActionFVariable) && i !== j){
+                    if(current_symbol.context === symbol_table.getSymbol(j).context && current_symbol.name === symbol_table.getSymbol(j).name){
+                        if(current_symbol.type === symbol_table.getSymbol(j).type){
                             continue;
                         }
                         else{
-                            build_error_display(symbol_table.getSymbol(i).node, editor, diags, "Error: inconsistent variables, variables with the same name in the same rule must have same types ");
+                            build_error_display(current_symbol.node, editor, diags, "Error: inconsistent variables, variables with the same name in the same rule must have same types ");
                             break;
                         }
                     }
@@ -131,13 +155,41 @@ function check_variables_type_is_consistent_inside_a_rule(symbol_table : Tamarin
                 }
             }
         }
+        else if (current_symbol.declaration === DeclarationType.REquationVariable){
+            let isbreak = false;
+            for (let symbol of symbol_table.getSymbols()){
+                if(symbol.declaration === DeclarationType.LEquationVariable && symbol.name === current_symbol.name && symbol.context === current_symbol.context){
+                    isbreak = true;
+                    break;
+                }
+            }
+            if (!isbreak){
+                build_error_display(current_symbol.node, editor, diags, "Error : this variable doesn't exist on the left side of the equation")
+            }
+        }
+        else if (current_symbol.declaration === DeclarationType.RMacroVariable){
+            let isbreak = false;
+            if(current_symbol.type === '$'){
+                continue;
+            }
+            for (let symbol of symbol_table.getSymbols()){
+                if(symbol.declaration === DeclarationType.LMacroVariable && symbol.name === current_symbol.name && symbol.context === current_symbol.context){
+                    isbreak = true;
+                    break;
+                }
+            }
+            if (!isbreak){
+                build_error_display(current_symbol.node, editor, diags, "Error : this variable doesn't exist on the left side of the equation")
+            }
+        }
         else{
             continue;
         }
     }
 };
 
-function check_case_sensitivity(symbol_table : TamarinSymbolTable, editor: vscode.TextEditor, diags: vscode.Diagnostic[]){
+/* Function used to check the speling of facts, also provides a quick fix for the wrong ones using leverstein distance */
+function check_case_sensitivity(symbol_table : TamarinSymbolTable, editor: vscode.TextEditor, diags: vscode.Diagnostic[]): void{
     const facts : TamarinSymbol["name"][]  = [];
     let count = 0;
     for( let i = 0 ; i < symbol_table.getSymbols().length; i++){
@@ -159,8 +211,6 @@ function check_case_sensitivity(symbol_table : TamarinSymbolTable, editor: vscod
                     continue;
                 }
                 //---------
-                let closestName = "";
-                let closestDistance = Number.MAX_VALUE;
                 for( let j = 0; j < facts.length ; j++ ){
                     const name2 = facts[j];
                     if(name2){
@@ -168,24 +218,19 @@ function check_case_sensitivity(symbol_table : TamarinSymbolTable, editor: vscod
                             continue;
                         }
                         const distance = levenshteinDistance(name, name2);
-                        if (distance < closestDistance) {
-                            closestName = name2;
-                            closestDistance = distance;
+                        if (distance < 3 && !facts.includes(current_symbol.name)) { // threshold value
+                            const diagnostic = build_warning_display(current_symbol.node, editor, diags, "Warning: did you mean " + name2 + " ? (" + distance + "characters away)")
+                            diagnostic.code = "wrongFactName";
+                            const range = current_symbol.name_range;
+                            const fix = new vscode.CodeAction("Replace with " + name2, vscode.CodeActionKind.QuickFix);
+                            fix.edit = new vscode.WorkspaceEdit();
+                            fix.edit.replace(editor.document.uri, range, name2);
+                            fix.diagnostics = [diagnostic];
+                            fix.isPreferred = true;
+                            fixMap.set(diagnostic, fix);
+                            count ++
                         }
                     }
-                }
-                if (closestDistance < 3 && !facts.includes(current_symbol.name)) { // threshold value
-                    const diagnostic = build_warning_display(current_symbol.node, editor, diags, "Warning: did you mean " + closestName + " ? (" + closestDistance + "characters away)")
-                    diagnostic.code = "wrongFactName";
-                    const range = current_symbol.name_range;
-                    const fix = new vscode.CodeAction("Replace with " + closestName, vscode.CodeActionKind.QuickFix);
-                    fix.edit = new vscode.WorkspaceEdit();
-                    fix.edit.replace(editor.document.uri, range, closestName);
-                    fix.diagnostics = [diagnostic];
-                    fix.isPreferred = true;
-                    fixMap.set(diagnostic, fix);
-                    count ++
-                    break;
                 }
             }
         }
@@ -195,20 +240,13 @@ function check_case_sensitivity(symbol_table : TamarinSymbolTable, editor: vscod
 
 
 
-
-
-
-
-
-
-
-
-
-function check_variable_is_defined_in_premise(symbol_table : TamarinSymbolTable, editor: vscode.TextEditor, diags: vscode.Diagnostic[]){
+/* Function used to check if a variable present in an action fact or conclusion is also present in premise,
+also checks if a fact in premise appears in a conclusion somewhere else */ 
+function check_variable_is_defined_in_premise(symbol_table : TamarinSymbolTable, editor: vscode.TextEditor, diags: vscode.Diagnostic[]):void{
     for( let i = 0 ; i < symbol_table.getSymbols().length; i++){
         let current_symbol = symbol_table.getSymbol(i);
         if(current_symbol.type === '$'){continue};  // Do not take into account public variables
-        if(current_symbol.declaration === DeclarationType.CCLVariable){
+        if(current_symbol.declaration === DeclarationType.CCLVariable || current_symbol.declaration === DeclarationType.ActionFVariable){
             let current_context = current_symbol.context;
             let is_break = false;
             for (let j = 0; j < symbol_table.getSymbols().length; j++){
@@ -225,7 +263,7 @@ function check_variable_is_defined_in_premise(symbol_table : TamarinSymbolTable,
                 }
             }
             if(!is_break){
-                build_error_display(current_symbol.node, editor, diags, "Error: this variable is used in conclusion but doesn't appear in premise");
+                build_error_display(current_symbol.node, editor, diags, "Error: this variable is used in the second part of the rule but doesn't appear in premise");
             }
         }
         else if ((current_symbol.declaration === DeclarationType.LinearF || current_symbol.declaration === DeclarationType.PersistentF) && current_symbol.node.parent?.grammarType === DeclarationType.Premise){
@@ -245,6 +283,7 @@ function check_variable_is_defined_in_premise(symbol_table : TamarinSymbolTable,
     }
 }
 
+/* This function performs various checks on action facts : wether they are declared ord not and if the arity is correct */ 
 function check_action_fact(symbol_table : TamarinSymbolTable, editor: vscode.TextEditor, diags: vscode.Diagnostic[]){
     let actionFacts: TamarinSymbol[] = [];
     let errors : string[] = []
@@ -283,6 +322,8 @@ function check_action_fact(symbol_table : TamarinSymbolTable, editor: vscode.Tex
     }
 }
 
+/* Function to check macros, functions, facts arity and to provide
+ quick fixes for wrong function name still using leverstein distance */
 function check_function_macros_and_facts_arity(symbol_table : TamarinSymbolTable, editor: vscode.TextEditor, diags: vscode.Diagnostic[]){
     let known_functions : TamarinSymbol[] = [];
     let errors : string[] = []
@@ -338,21 +379,36 @@ function check_function_macros_and_facts_arity(symbol_table : TamarinSymbolTable
         }
     }
     for(let symbol of known_functions){
-        let isbreak = false
+        let isbreak = false;
         if(symbol.declaration === DeclarationType.NARY){
-            for( let functions of known_functions){
-                if(functions.name === symbol.name && functions !== symbol){
+            for( let functionSymbol of known_functions){
+                if(functionSymbol.name === symbol.name && functionSymbol !== symbol){
                     isbreak = true;
                     break;
                 }
+                if (typeof symbol.name === 'string' && typeof functionSymbol.name === 'string' && symbol.name !== functionSymbol.name && symbol.arity === functionSymbol.arity) {
+                    const distance = levenshteinDistance(symbol.name, functionSymbol.name);
+                    if (distance < 3) { // threshold value
+                        const diagnostic = build_warning_display(symbol.node, editor, diags, "Warning: did you mean " + functionSymbol.name + " ? (" + distance + "characters away)");
+                        diagnostic.code = "wrongFunctionName";
+                        const range = symbol.name_range;
+                        const fix = new vscode.CodeAction("Replace with " + functionSymbol.name, vscode.CodeActionKind.QuickFix);
+                        fix.edit = new vscode.WorkspaceEdit();
+                        fix.edit.replace(editor.document.uri, range, functionSymbol.name);
+                        fix.diagnostics = [diagnostic];
+                        fix.isPreferred = true;
+                        fixMap.set(diagnostic, fix);
+                    }
+                }
             }
             if(!isbreak){
-                build_error_display(symbol.node, editor, diags, "Error : unknown function or macro")
+                build_error_display(symbol.node, editor, diags, "Error : unknown function or macro");
             }
         }
     }
-}
+}    
 
+/* Function used to check if a term is associated to qa quantified formula in a lemma */ 
 function check_free_term_in_lemma(symbol_table : TamarinSymbolTable, editor: vscode.TextEditor, diags: vscode.Diagnostic[]){
     let lemma_vars : TamarinSymbol[] = [];
     for (let symbol of symbol_table.getSymbols()){
@@ -384,7 +440,8 @@ function check_free_term_in_lemma(symbol_table : TamarinSymbolTable, editor: vsc
                     gt_list = get_child_grammar_type(search_context);
                     } 
                 }
-                if(search_context.grammarType === DeclarationType.Lemma || search_context.grammarType === DeclarationType.Restriction){
+                // Simultaneously defining context for rename feature
+                if(search_context.grammarType === DeclarationType.Lemma || search_context.grammarType === DeclarationType.Restriction || search_context.grammarType === 'diff_lemma'){
                     set_associated_qf(lemma_vars[i], search_context.child(4));
                 }
                 else if(search_context.grammarType === DeclarationType.NF){
@@ -418,29 +475,159 @@ function check_free_term_in_lemma(symbol_table : TamarinSymbolTable, editor: vsc
     }
 }
 
+/* Function to check that a macro is not used in an equation */ 
+function check_macro_not_in_equation(symbol_table : TamarinSymbolTable, editor: vscode.TextEditor, diags: vscode.Diagnostic[]){
+    for(let symbol of symbol_table.getSymbols()){
+        if(symbol.declaration === DeclarationType.NARY){
+            for ( let macros of symbol_table.getSymbols()){
+                if(macros.declaration === DeclarationType.Macro && macros.name === symbol.name && symbol.context.grammarType === DeclarationType.Equation){
+                    build_error_display(symbol.node, editor, diags, "Error : a macro shoud not be used in an equation ")
+                }
+            }
+        }
+    }
+}
 
+/* Given a symbol table returns all the builtins present in it */ 
+function return_builtins(symbol_table: TamarinSymbolTable): TamarinSymbol[]{
+let builtins : TamarinSymbol[]  = [];
+    for (let symbol of symbol_table.getSymbols()){
+        if(symbol.declaration === DeclarationType.Builtin){
+            builtins.push(symbol);
+        }
+    } 
+    return builtins;
+}
+
+/* Given a list of builtins returns there name */
+function get_builtins_name(builtins : TamarinSymbol[]): string[]{
+    let Sbuiltins : string[] = [];
+    for (let builtin of builtins ){
+        if(builtin.name)
+        Sbuiltins.push(builtin.name)
+    }
+    return Sbuiltins;
+}
+
+function return_functions(symbol_table: TamarinSymbolTable): string[]{
+    let builtins : string[]  = [];
+        for (let symbol of symbol_table.getSymbols()){
+            if(symbol.declaration === DeclarationType.Functions && symbol.name){
+                builtins.push(symbol.name);
+            }
+        } 
+        return builtins;
+    }
+
+/* Function used to check if the use of * ^ or others symbol is allowed, if not provides a quick fix to include the right builtin,
+Also works with functions defined in builtins*/ 
+function check_infix_operators(symbol_table : TamarinSymbolTable, editor : vscode.TextEditor, diags : vscode.Diagnostic[], root : Parser.SyntaxNode){
+
+    function display_infix_error(builtin: string, symbol: string, child: Parser.SyntaxNode): void {
+        let current_builtins = return_builtins(symbol_table);
+        let current_functions = return_functions(symbol_table)
+        if (!get_builtins_name(current_builtins).includes(builtin) && !current_functions.includes(getName(child.child(0), editor))) {
+            const diagnostic = build_error_display(child, editor, diags, "Error : symbol " + symbol + " cannot be used without " + builtin + " builtin");
+            diagnostic.code = "missingBuiltin";
+            const fix = new vscode.CodeAction("Include builtin : " + builtin, vscode.CodeActionKind.QuickFix);
+            if(current_builtins.length > 0){
+            const range = current_builtins[current_builtins.length - 1].name_range; 
+            fix.edit = new vscode.WorkspaceEdit();
+            fix.edit.insert(editor.document.uri, range.end, ", " + builtin); 
+            }
+            else {
+                let theory = root ;
+                while (theory.grammarType !== 'theory'){
+                    if(theory.parent)
+                    theory = theory.parent
+                }
+                theory = theory.child(3) as Parser.SyntaxNode
+                const range = new vscode.Range(
+                    editor.document.positionAt(theory.startIndex),
+                    editor.document.positionAt(theory.endIndex)
+                );
+                fix.edit = new vscode.WorkspaceEdit();
+                fix.edit.insert(editor.document.uri, range.end, "\nbuiltins :  " + builtin);
+            }
+            fix.diagnostics = [diagnostic];
+            fix.isPreferred = true;
+            fixMap.set(diagnostic, fix);     
+        }
+      }
+      
+    for (let child of root.children){
+        if(child.grammarType === '^' || child.grammarType === '*'){
+            let current_builtins = return_builtins(symbol_table);
+            if(! get_builtins_name(current_builtins).includes('diffie-hellman')){
+                display_infix_error('diffie-hellman', '^ or *', child) 
+                }
+        }
+        else if (child.grammarType === '⊕'){
+            display_infix_error('xor','⊕', child);;
+        }
+        else if (child.grammarType === '++'){
+            display_infix_error('multiset', '++', child);;
+        }
+        else if (child.grammarType === '%+'){
+            display_infix_error('natural-numbers', '%+', child);
+        }
+        else if (child.grammarType === DeclarationType.NARY){
+            if(getName(child.child(0),editor) === 'inv'){
+                display_infix_error('diffie-hellman', 'inv', child);
+            }
+            else if (getName(child.child(0),editor) === 'h'){
+                display_infix_error('hashing', 'h', child);
+            }
+            else if (getName(child.child(0), editor) === 'sdec' || getName(child.child(0), editor) === 'senc'){
+                display_infix_error('symmetric-encryption', 'sdec or senc' , child);
+            }
+            else if (getName(child.child(0), editor) === 'adec' || getName(child.child(0), editor) === 'aenc'){
+                display_infix_error('asymmetric-encryption', 'adec or aenc' , child);
+            }
+            else if (getName(child.child(0), editor) === 'sign' || getName(child.child(0), editor) === 'verify'){
+                display_infix_error('signing', 'sign or verify' , child);
+            }
+            else if (getName(child.child(0), editor) === 'revealSign' || getName(child.child(0), editor) === 'revealVerify' || getName(child.child(0), editor) === 'getMessage'){
+                display_infix_error('revealing-signing', 'revealSign or revealVerify or getMessage' , child);
+            }
+            else if (getName(child.child(0), editor) === 'pmult' || getName(child.child(0), editor) === 'em'){
+                display_infix_error('bilinear-pairing', 'pmult or em' , child);
+            }
+            else if (getName(child.child(0),editor) === 'XOR'){
+                display_infix_error('xor', 'XOR', child);  
+            }
+            
+        }
+        else (check_infix_operators(symbol_table,editor,diags,child));
+    }
+
+
+}
 
 const fixMap = new Map<vscode.Diagnostic, vscode.CodeAction>();
 
+// Register the code action provider usefull for the quick fixes 
 vscode.languages.registerCodeActionsProvider('tamarin', {
-    provideCodeActions(document: vscode.TextDocument, range: vscode.Range, context: vscode.CodeActionContext, token: vscode.CancellationToken) {
-        const actions: vscode.CodeAction[] = [];
-        for (const diagnostic of context.diagnostics) {
-            const fix = fixMap.get(diagnostic);
-            if (fix) {
-                actions.push(fix);
-            }
-        }
-        return actions;
+  provideCodeActions(document: vscode.TextDocument, range: vscode.Range, context: vscode.CodeActionContext, token: vscode.CancellationToken) {
+    const actions: vscode.CodeAction[] = [];
+    for (const diagnostic of context.diagnostics) {
+      const fix = fixMap.get(diagnostic);
+      if (fix) {
+        actions.push(fix);
+      }
     }
+    return actions;
+  }
 });
 
-
-export function checks_with_table(symbol_table : TamarinSymbolTable, editor: vscode.TextEditor, diags: vscode.Diagnostic[]){
+export function checks_with_table(symbol_table : TamarinSymbolTable, editor: vscode.TextEditor, diags: vscode.Diagnostic[], root : Parser.SyntaxNode){
     check_variables_type_is_consistent_inside_a_rule(symbol_table, editor, diags);
-    check_case_sensitivity(symbol_table, editor, diags);
     check_variable_is_defined_in_premise(symbol_table, editor, diags);
     check_action_fact(symbol_table, editor, diags);
     check_function_macros_and_facts_arity(symbol_table, editor, diags);
     check_free_term_in_lemma(symbol_table, editor, diags);
+    check_macro_not_in_equation(symbol_table, editor, diags)
+    check_infix_operators(symbol_table, editor, diags, root);
+    check_case_sensitivity(symbol_table, editor, diags);
 };
+